Minimized string copies from AssetIDs

+++ src/common/assets/types/materialAsset.h
@@ -11,7 +11,7 @@ public:
     AssetID vertexShader;
     AssetID fragmentShader;
     AssetID inputComponent;
-    size_t runtimeID = -1;
+    uint32_t runtimeID = -1;
 
     void serialize(OutputSerializer& s) const override;
     void deserialize(InputSerializer& s) override;
